btier: fix build warning
drivers/block/btier/btier_main.c: In function 'tier_ioctl':
drivers/block/btier/btier_main.c:2419:33: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

+++ drivers/block/btier/btier_main.c
@@ -2416,7 +2416,7 @@ static long tier_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
 			if (0 != (err = order_devices(dev)))
 				break;
 			if (0 == (err = determine_device_size(dev))) {
-                                pr_err("dev->users at %llu\n",(u64)&dev->users);
+				pr_err("dev->users at %lu\n",(uintptr_t)&dev->users);
 				err = tier_register(dev);
 			}
 		}
